backend/vector_search_mistral/requirements.txt
# Core dependencies
mistralai>=0.0.7
pinecone>=6.0.0
# pinecone-client>=6.0.0  # Old SDK name
openai>=1.0.0  # OpenAI API for enhanced RAG capabilities

# PDF processing
# PyPDF2>=1.26.0  # Uncomment if needed for basic PDF parsing
# pdf2image>=1.12.1  # Uncomment if OCR is needed
# pytesseract>=0.3.7  # Uncomment if OCR is needed

# Environment and utilities
python-dotenv>=1.0.0
numpy>=1.24.0
nltk>=3.8.1
tiktoken>=0.5.1

# Web and networking
# flask>=2.0.0  # Uncomment if web server is needed
requests>=2.31.0

# File and path handling
pathlib>=1.0.1 